+++ platform/viewer/src/context/WhiteLabellingContext.js
@@ -1,4 +1,4 @@
-import OHIFLogo from './components/OHIFLogo/OHIFLogo.js';
+import OHIFLogo from '../components/OHIFLogo/OHIFLogo.js';
 import React from 'react';
 
 const defaultContextValues = {